Rename and split files
